# reduce bytes used when serializing Extent