This commit reflects comments made by Adrien in https://github.com/elastic/elasticsearch/pull/50834 surrounding the Extent serialization.

it re-orders and negates a few values in order to save more space