@talevy 2020-02-20
@jpountz thanks for the suggestions.

I did a rough speed/size test of these changes and found the following
performance numbers on 3 million random Extents. recorded time spent 
writing/reading an Extent and its size.

before changes:

writeSpeed: 172.679105 ns
readSpeed: 53.29868 ns
byte-size: 19.23683 bytes

after changes:

writeSpeed: 169.303268 ns
readSpeed: 52.649607 ns
byte-size: 19.163605 bytes